backend/src/store_consumer/requirements.txt
python-dotenv==0.19.0
kafka-python==2.0.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9  # Changed from psycopg2 to psycopg2-binary